requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
pydantic==1.8.2
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
djangorestframework~=3.14.0